requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP Client
httpx>=0.26.0

# Vector Store
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.3.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# Utilities
python-dotenv>=1.0.0
